Add files via upload
This Python project uses the "pyshorteners" library to shorten URLs. This allows users to enter a long URL and get a shortened version using services provided by "pyshorteners". The script prompts the user for a URL, processes it, and returns a shortened version. Choosing a specific service for shortening can be customized by changing the method in the code "shortener.tinyurl.short".